load-simulator/requirements.txt
psycopg2-binary==2.9.9
prometheus-client==0.17.1
Faker==19.3.1
